Add companion bubble, merge WebSocket, delete CompanionScreen
- Expand onStepComplete: add show_pose detection + tool failure logging
- Add onSpeakReady/onSpeakError callbacks (TTS integration)
- Bind sendChatMessage voiceEnabled to companionOpen && voiceEnabled
- Add stopSpeaking cleanup on companion close, agent switch, unmount
- Add companion reply bubble with FadeInDown animation above FIB
- Bubble shows latest AI message, 'thinking...' placeholder, ScrollView
- Delete CompanionScreen.tsx (all features migrated to ChatScreen)
